Move the display storage manifest models to the display lib
