+++ display/src/main/scala/uk/ac/wellcome/platform/archive/display/manifests/DisplayFileManifest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.storage.bags.api.models
+package uk.ac.wellcome.platform.archive.display.manifests
 
 import io.circe.generic.extras.JsonKey
 import uk.ac.wellcome.platform.archive.common.storage.models.FileManifest
